// src/app/appmanager/model/component.ts
/**
 * Kathra Core Model
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0-RC-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ApiVersion } from './apiVersion';
import { Implementation } from './implementation';
import { Library } from './library';
import { Resource } from './resource';
import { SourceRepository } from './sourceRepository';


export interface Component {
    id?: string;
    name?: string;
    status?: Component.StatusEnum;
    createdBy?: string;
    updatedBy?: string;
    createdAt?: number;
    updatedAt?: number;
    metadata?: { [key: string]: any; };
    description?: string;
    title?: string;
    versions?: Array<ApiVersion>;
    libraries?: Array<Library>;
    implementations?: Array<Implementation>;
    apiRepository?: SourceRepository;
}
export namespace Component {
    export type StatusEnum = 'PENDING' | 'READY' | 'UPDATING' | 'UNSTABLE' | 'ERROR' | 'DELETED';
    export const StatusEnum = {
        PENDING: 'PENDING' as StatusEnum,
        READY: 'READY' as StatusEnum,
        UPDATING: 'UPDATING' as StatusEnum,
        UNSTABLE: 'UNSTABLE' as StatusEnum,
        ERROR: 'ERROR' as StatusEnum,
        DELETED: 'DELETED' as StatusEnum
    }
}